requirements.txt
asgiref==3.8.1
astroid==3.2.4
dill==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
isort==5.13.2
Markdown==3.7
mccabe==0.7.0
packaging==24.1
platformdirs==4.3.3
pylint==3.2.7
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
uritemplate==4.1.1
